Moved items to one folder, extracted Mongobibs and used them samplewise
