+++ Medusa/ImageGenerator.py
