+++ Medusa/MedusaMongo.py
@@ -1,0 +1,28 @@
+import pymongo 
+
+################# Mongo Helpers #######################
+
+uri = "mongodb://MedusaUser:P3R5EU?@applis.me/Medusa?authSource=Medusa"
+
+def getMedusaTrainingCollection():
+    mongoClient=pymongo.MongoClient(uri)
+    medusaDB = mongoClient["Medusa"]
+    TrainingCollection = medusaDB["StratifiedTraining"]
+    return TrainingCollection
+
+def getMedusaTestCollection():
+    mongoClient=pymongo.MongoClient(uri)
+    medusaDB = mongoClient["Medusa"]
+    TestCollection = medusaDB["StratifiedTest"]
+    return TestCollection
+
+def get_Image_cursor(collection):
+    cursor = collection.find()
+    return cursor
+
+def get_next_n_samples(cursor,n = 100):
+    samples = []
+    for _ in range(n):
+        entry = cursor.next()
+        samples.append((entry["success"],entry["image"]))
+    return samples
